# Make Range representation lightweight