Previously `Range` consisted of low and high `Marker` objects, each
containing a single-row `Block`. This was useful from serialization
perspective, but caused:

1. memory overhead when dealing with many `Range` instances at once
2. API overhead, where consumer of `Range` had to translate `Marker`
   bounds to infer low-unboundedness and high-unboundedness of a range.

Now that `Range` is no longer serialized, we can simplify the
representation.


https://jmh.morethan.io/?sources=https://gist.githubusercontent.com/findepi/b082f654a9be4378a4403688eed2504e/raw/51e198fee3058d8c925898d48e075048eecd13b6/JMH%2520BenchmarkSortedRangeSet%2520A1%2520master.json,https://gist.githubusercontent.com/findepi/b082f654a9be4378a4403688eed2504e/raw/51e198fee3058d8c925898d48e075048eecd13b6/JMH%2520BenchmarkSortedRangeSet%2520A2%2520branch.json

#### before

```
Benchmark                                      Mode  Cnt     Score     Error  Units
BenchmarkSortedRangeSet.benchmarkBuilder       avgt   10     0.986 ±   0.447  ms/op
BenchmarkSortedRangeSet.complementLarge        avgt   10    36.320 ±   1.351  ms/op
BenchmarkSortedRangeSet.complementSmall        avgt   10   238.683 ±  26.884  ms/op
BenchmarkSortedRangeSet.containsValueLarge     avgt   10  1115.749 ± 130.025  ms/op
BenchmarkSortedRangeSet.containsValueSmall     avgt   10   550.078 ±  40.958  ms/op
BenchmarkSortedRangeSet.equalsLarge            avgt   10     0.057 ±   0.003  ms/op
BenchmarkSortedRangeSet.equalsSmall            avgt   10    14.319 ±   3.276  ms/op
BenchmarkSortedRangeSet.getOrderedRangesLarge  avgt   10   927.169 ±   8.283  ms/op
BenchmarkSortedRangeSet.getOrderedRangesSmall  avgt   10   715.147 ±  30.282  ms/op
BenchmarkSortedRangeSet.overlapsLarge          avgt   10     8.690 ±   0.249  ms/op
BenchmarkSortedRangeSet.overlapsSmall          avgt   10   123.784 ±  13.068  ms/op
BenchmarkSortedRangeSet.unionLarge             avgt   10   269.696 ±  31.384  ms/op
BenchmarkSortedRangeSet.unionSmall             avgt   10   417.966 ± 111.192  ms/op
```


#### after

```
Benchmark                                      Mode  Cnt     Score     Error  Units
BenchmarkSortedRangeSet.benchmarkBuilder       avgt   10     1.112 ±   0.233  ms/op
BenchmarkSortedRangeSet.complementLarge        avgt   10    36.034 ±   1.859  ms/op
BenchmarkSortedRangeSet.complementSmall        avgt   10   226.923 ±  19.067  ms/op
BenchmarkSortedRangeSet.containsValueLarge     avgt   10  1064.851 ±  35.166  ms/op
BenchmarkSortedRangeSet.containsValueSmall     avgt   10   555.789 ±  43.273  ms/op
BenchmarkSortedRangeSet.equalsLarge            avgt   10     0.058 ±   0.005  ms/op
BenchmarkSortedRangeSet.equalsSmall            avgt   10    14.006 ±   4.968  ms/op
BenchmarkSortedRangeSet.getOrderedRangesLarge  avgt   10   187.961 ±   7.599  ms/op
BenchmarkSortedRangeSet.getOrderedRangesSmall  avgt   10   206.972 ±  88.700  ms/op
BenchmarkSortedRangeSet.overlapsLarge          avgt   10     5.255 ±   0.156  ms/op
BenchmarkSortedRangeSet.overlapsSmall          avgt   10    78.558 ±  32.456  ms/op
BenchmarkSortedRangeSet.unionLarge             avgt   10   204.506 ±  50.603  ms/op
BenchmarkSortedRangeSet.unionSmall             avgt   10   360.224 ± 133.375  ms/op
```